+++ CHANGELOG.md
@@ -5,6 +5,14 @@
 * RiverDriftDetectionCondition for wrapping the river drift detection algorithms ([#221](https://github.com/KIT-IAI/pyWATTS/issues/221))
 
 
+
+
+## 0.X.X - Unreleased
+
+### Added
+* Add a module for statistic feature extraction ([#217](https://github.com/KIT-IAI/pyWATTS/issues/217))
+
+
 ## 0.3.0 - 2022-31-03
 
 ### Added